user added

+++ src/main/java/ru/anikanov/tm/service/ProjectService.java
@@ -3,6 +3,7 @@ package ru.anikanov.tm.service;
 import ru.anikanov.tm.entity.Project;
 import ru.anikanov.tm.repository.ProjectRepository;
 import ru.anikanov.tm.repository.TaskRepository;
+import ru.anikanov.tm.repository.UserRepository;
 
 import java.text.ParseException;
 import java.util.ArrayList;
@@ -11,10 +12,12 @@ import java.util.List;
 public class ProjectService {
     private ProjectRepository projectRepository;
     private TaskRepository taskRepository;
+    private UserRepository userRepository;
 
-    public ProjectService(ProjectRepository pr, TaskRepository tr) {
+    public ProjectService(ProjectRepository pr, TaskRepository tr, UserRepository ur) {
         projectRepository = pr;
         taskRepository = tr;
+        userRepository = ur;
     }
 
     public Project persist(String projectName, String description, String dateStart, String dateFinish, String userId) throws ParseException {
@@ -30,9 +33,11 @@ public class ProjectService {
         return null;
     }
 
-    public void merge(String projectName, String description, String dateStart, String dateFinish) throws ParseException {
+    public void merge(String projectName, String description, String dateStart, String dateFinish, String userId) throws ParseException {
         if (projectName.isEmpty() || (projectName == null)) return;
         Project project = projectRepository.findOne(projectName);
+        if ((!userId.equals(project.getUserId())) || (!userRepository.findOne(userId).getRole().equals("admin")))
+            return;
         if (project != null) {
             if (description.isEmpty() || (description == null)) return;
             if (dateStart.isEmpty() || (dateStart == null)) return;
@@ -41,9 +46,11 @@ public class ProjectService {
         projectRepository.merge(projectName, description, dateStart, dateFinish);
     }
 
-    public void remove(String projectName) {
+    public void remove(String projectName, String userId) {
         if (projectName.isEmpty() || (projectName == null)) return;
         Project project = projectRepository.findOne(projectName);
+        if ((!userId.equals(project.getUserId())) || (!userRepository.findOne(userId).getRole().equals("admin")))
+            return;
         if (project != null) {
             String projectId = project.getId();
             projectRepository.remove(projectName);
@@ -51,16 +58,18 @@ public class ProjectService {
         }
     }
 
-    public void removeAll() {
+    public void removeAll(String userId) {
+        if (!userRepository.findOne(userId).getRole().equals("admin")) return;
         projectRepository.removeAll();
         taskRepository.removeAll();
     }
 
-    public List<Project> findAll() {
+    public List<Project> findAll(String userId) {
+        if (!userRepository.findOne(userId).getRole().equals("admin")) return null;
         return projectRepository.findAll();
     }
 
-    public Project findOne(String projectName) {
+    public Project findOne(String projectName, String userId) {
         if (projectName.isEmpty() || (projectName == null)) return null;
         return projectRepository.findOne(projectName);
     }
